More work on deduplicating notification
